Moved board printing into its own module
